Add chapters relation to course (#129)
- Added chapters as one to many relations in the course.
- Now chapters of the course can be accessed by course.getChapters()

+++ course/src/test/java/in/testpress/course/ui/ChaptersListFragmentTest.java
@@ -51,7 +51,7 @@ public class ChaptersListFragmentTest {
         chapter.setSlug("chapter");
         chapter.setName("Chapter");
         chapter.setContentsCount(3);
-        chapter.setCourseId(1);
+        chapter.setCourseId((long)1);
         chapter.setActive(true);
         chapterDao.insertInTx(chapter);
         this.chapter = chapter;
